Adding in skeleton.
This adds in days and weeks for the entire semester. They're generated
by a script, which will be added in momentarily.

I don't know if this is the best way to do this, but it doesn't seem
bad. It allows me to include a week anywhere I want very easily... in
particular, on the homepage.

An "archive" is just a matter of having a list of one-liners on a
separate page, and linking should be easy enough. So, net-net, it looks
like a win.